Move spawn_multiprocess into UnprivilegedPipelineContent.
